% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_wrappers.R
\name{generate_diverse_demographics}
\alias{generate_diverse_demographics}
\title{Generate Diverse Demographics}
\usage{
generate_diverse_demographics(n)
}
\arguments{
\item{n}{Integer number of profiles to generate}
}
\value{
Data frame with demographic information
}
\description{
Internal function to generate diverse demographic profiles for participants.
}
\keyword{internal}
