% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_wrappers.R
\name{generate_survey_responses}
\alias{generate_survey_responses}
\title{Generate Survey Responses}
\usage{
generate_survey_responses(n)
}
\arguments{
\item{n}{Integer number of response sets to generate}
}
\value{
Data frame with survey responses
}
\description{
Internal function to generate diverse survey responses for participants.
}
\keyword{internal}
