% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FocusGroup.R
\name{print.FocusGroup}
\alias{print.FocusGroup}
\title{Print a FocusGroup}
\usage{
\method{print}{FocusGroup}(x, ...)
}
\arguments{
\item{x}{A `FocusGroup` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a FocusGroup
}
\keyword{internal}
