% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_wrappers.R
\name{generate_persona}
\alias{generate_persona}
\title{Generate Persona}
\usage{
generate_persona(demographics, survey_responses = NULL)
}
\arguments{
\item{demographics}{List of demographic information}

\item{survey_responses}{List of survey responses (optional)}
}
\value{
Character string with persona description
}
\description{
Internal function to generate a persona description based on demographics and survey responses.
}
\keyword{internal}
